+++ src/pyAdvanced_license/pyAdvanced_license/fingerprint.py
